Add trycompile, a prototype functioning like trybuild but using a test environment
Trybuild tests whether specific rust files compile or not, which is especially useful for macro development. However, it uses an environment where #[cfg(test)] is false.
Trycompile is a quick prototype inspired by Trybuild which uses `cargo test --no-run` as to build crates in a #[cfg(test)] is true environment.

We should aim to open a PR (if the trybuild author agrees of course!) to trybuild which enables running in an environment with #[cfg(test)], or figure out how to enable it (perhaps a RUSTFLAG?), as trybuild is well tested and has not been created in an hour or so (i.e. has been developed with proper thought, instead of quickly hacking something together).