Mark windows as unsupported for trycompile
If I will develop (and likely even if not) trycompile further, I will look into adding windows support. For now however I marked it as unsupported. Tests run on windows will always succeed, but no actual test case will run. This ensures CI succeeds when testing on the windows platform (while a unix platform is still be able to fail, since there the tests will be run regularly). The reason for adding this is that sif_macro also uses this test runner now as a complementary runner to trybuild, but the CI fails on windows. I would have added a feature in the Cargo.toml, and let the test table for trycompile have that required-feature, but this isn't possible since Cargo doesn't support optional features for dev-dependencies.